refactor: flatten rollup test files
