+++ packages/rollup/test/results/no-maps.css
@@ -1,0 +1,4 @@
+/* packages/rollup/test/specimens/simple.css */
+.fooga {
+    color: red
+}
