--- packages/rollup/test/results/rollup/no-maps.css
@@ -1,4 +1,0 @@
-/* packages/rollup/test/specimens/rollup/simple.css */
-.fooga {
-    color: red
-}
