--- packages/rollup/test/results/rollup/simple.css
@@ -1,5 +1,0 @@
-/* packages/rollup/test/specimens/rollup/simple.css */
-.fooga {
-    color: red
-}
-/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIi4uLy4uL3NwZWNpbWVucy9yb2xsdXAvc2ltcGxlLmNzcyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiQUFBQSxzREFBQztBQUFEO0lBQ0ksVUFBVztDQUNkIiwiZmlsZSI6InNpbXBsZS5jc3MiLCJzb3VyY2VzQ29udGVudCI6WyIuZm9vZ2Ege1xuICAgIGNvbG9yOiByZWQ7XG59XG4iXX0= */
