--- packages/rollup/test/results/rollup/simple.json
@@ -1,5 +1,0 @@
-{
-    "packages/rollup/test/specimens/rollup/simple.css": {
-        "fooga": "fooga"
-    }
-}
