+++ packages/rollup/test/specimens/dependencies.js
