+++ packages/rollup/test/specimens/invalid-name.js
