+++ packages/rollup/test/specimens/tree-shaking.js
